+++ package.json
@@ -89,7 +89,6 @@
     "react-accessible-treeview": "^2.6.2",
     "react-bootstrap": "^0.31.5",
     "react-breadcrumbs": "^2.1.6",
-    "react-burger-menu": "^2.6.10",
     "react-chartjs-2": "^5.2.0",
     "react-data-export": "^0.5.0",
     "react-datetime": "^2.15.0",
