+++ src/styles/landing.css
@@ -52,6 +52,10 @@ body {
 
 .primary-layout {
   margin: 30px 0;
+
+  @media (min-width: 768px) {
+    margin: 30px 0 30px 40px;
+  }
 }
 
 .container {
